% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_make.R
\name{all_make}
\alias{all_make}
\title{初期状態からplot直前までの加工全部}
\usage{
all_make(samples, coneN, height, azimuth, control)
}
\arguments{
\item{samples}{標本点}

\item{coneN}{錐の数}

\item{height}{錐の大きさ}

\item{azimuth}{錐の最小の角度}

\item{control}{追加オプション}
}
\description{
初期状態からplot直前までの加工全部
}
\examples{
samples <- data.frame(x = rnorm(10), y = rnorm(10))
dat <- all_make(samples, 100, 1, pi/2, control = list(max_loop = 1e+05))
}

