% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{C2}
\alias{C2}
\title{Title C2}
\usage{
C2(x, y, feature_selection_method, num_clusters_method, k = NULL,
  clustering_method, ...)
}
\arguments{
\item{x}{data matrix}

\item{y}{Dependent variable}

\item{feature_selection_method}{method for the feature selection of the clinical measurements stage. Default RF.}

\item{num_clusters_method}{method for the choosing number of clusters by using the clinical measurements. Default Euclidean.}

\item{k}{number of clusters to use. If missing, we use a detection method. Defaukt as NULL}

\item{clustering_method}{method for clustering using the reduced clinical measures. Default is Hmanhattan,}
}
\value{
a list of three variables: 
1) vector with the names of the omportant variables chosen.
2) number of classes that will be used for clustering 
3) vector of the new assigned clusterst
}
\description{
Title C2
}
\examples{
resultC2 <- C2(x, y, feature_selection_method='RF', num_clusters_method='Manhattan', clustering_method='Manhattan', plot.num.clus=TRUE, plot.clustering=TRUE)
C2(x, y, feature_selection_method='BIC', num_clusters_method='Manhattan', clustering_method='Hmanhattan', plot.num.clus=TRUE, plot.clustering=FALSE, nbest=1, nvmax=8, B=50)
}

