% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{C3}
\alias{C3}
\title{Title C3}
\usage{
C3(PBx, newy, feature_selection_method, classification_method)
}
\arguments{
\item{PBx}{data matrix}

\item{newy}{new assigned clusters, results from C2.}

\item{feature_selection_method}{method for the feature selection of the Potential Bio-Markers}

\item{classification_method}{method for classification  using the potential bio-markers}
}
\value{
a list of two variables: 
1) vector with the names of important variables chosen 
2) classification result for each observation
}
\description{
Title C3
}
\examples{
C3(PBx, newy, feature_selection_method='RF', classification_method='RF') 

}

