% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc-package.R
\docType{package}
\name{ccc-package}
\alias{ccc}
\alias{ccc-package}
\title{ccc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: andrés castro araújo \email{acastroaraujo@gmai.com}

}
\keyword{internal}
