% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{ccc_clean_dataset}
\alias{ccc_clean_dataset}
\title{Clean Dataset}
\usage{
ccc_clean_dataset(df)
}
\arguments{
\item{df}{a data frame, as created by ccc_search()}
}
\value{
a data frame
}
\description{
Clean Dataset
}
