% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer-cosas.R
\name{ccc_elementos}
\alias{ccc_elementos}
\title{Extractor de expresiones en mayúscula en el preámbulo a cada sentencia}
\usage{
ccc_elementos(texto)
}
\arguments{
\item{texto}{}
}
\value{

}
\description{
Extractor de expresiones en mayúscula en el preámbulo a cada sentencia
}
