% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer-cosas.R
\name{ccc_fecha}
\alias{ccc_fecha}
\title{Extraer fecha}
\usage{
ccc_fecha(sentencia, texto)
}
\arguments{
\item{sentencia}{el nombre de la sentencia (e.g. C-776-03)}

\item{texto}{el texto de la sentencia}
}
\value{
a Date object
}
\description{
Extraer fecha
}
