% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer-cosas.R
\name{ccc_leyes_citadas}
\alias{ccc_leyes_citadas}
\title{Extractor de leyes citadas}
\usage{
ccc_leyes_citadas(texto)
}
\arguments{
\item{texto}{el texto de una sentencia, tal y como viene de ccc_extraer_texto()}
}
\value{
una lista de leyes citadas en el texto
}
\description{
Extractor de leyes citadas
}
