% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer-cosas.R
\name{ccc_magistrados}
\alias{ccc_magistrados}
\title{Información sobre magistrados}
\usage{
ccc_magistrados(texto)
}
\arguments{
\item{texto}{el texto de la sentencia}
}
\value{
a tibble with names and information for each Justice
}
\description{
Información sobre magistrados
}
