% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_url.R
\name{ccc_num_url}
\alias{ccc_num_url}
\title{Buscador de URL por número de providencia}
\usage{
ccc_num_url(q, p = 0)
}
\arguments{
\item{q}{términos de búsqueda}

\item{p}{número de página (default es cero)}
}
\value{
un véctor tipo character
}
\description{
La página de la Corte obliga a buscar en formatos como: T-005-04 o T-005; C-143-15; SU230-15 o SU230; A180-15
Pero también sirve buscar SU-230-15 o A-180-15 (con el guión agregado).
}
\details{
No hay restricción para el tipo de búsqueda (ej. Buscar 776 arroja todas las providencias que contienen ese número en el nombre)
}
