% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palabra-clave.R
\name{ccc_palabra_clave}
\alias{ccc_palabra_clave}
\title{Buscador de sentencias por palabra clave}
\usage{
ccc_palabra_clave(q, p = 0)
}
\arguments{
\item{q}{términos de búsqueda}

\item{p}{número de página (default es cero)}
}
\value{
un data frame
}
\description{
Buscador de sentencias por palabra clave
}
