% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palabra-clave.R
\name{ccc_palabra_clave2}
\alias{ccc_palabra_clave2}
\title{Buscador de sentencias por palabra clave (Nueva página)}
\usage{
ccc_palabra_clave2(q, p = 0, orden = c("relevancia", "fecha"), y = NULL)
}
\arguments{
\item{q}{términos de búsqueda}

\item{p}{número de página (default es cero)}

\item{orden}{por relevancia o por fecha}

\item{y}{año (default es "Todos los años")}
}
\value{
un data frame
}
\description{
Buscador de sentencias por palabra clave (Nueva página)
}
