% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer-texto.R
\name{ccc_pp}
\alias{ccc_pp}
\title{Extractor de pie de p치ginas}
\usage{
ccc_pp(path)
}
\arguments{
\item{path}{la direcci칩n de la sentencia (e.g. "/relatoria/2006/C-355-06.htm")}
}
\value{
un vector de pies de p치gina
}
\description{
Extractor de pie de p치ginas
}
