% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{ccc_search}
\alias{ccc_search}
\title{Buscador de Sentencias}
\usage{
ccc_search(text, date_start, date_end)
}
\arguments{
\item{text}{a text query}

\item{date_start}{yyyy/mm/dd}

\item{date_end}{yyyy/mm/dd}
}
\value{
a data frame
}
\description{
Buscador de Sentencias
}
