% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_cosas.R
\name{ccc_sentencias_citadas}
\alias{ccc_sentencias_citadas}
\title{Extractor de sentencias citadas}
\usage{
ccc_sentencias_citadas(texto)
}
\arguments{
\item{texto}{el texto de una sentencia, tal y como viene de ccc_extraer_texto()}
}
\value{
una lista de sentencias citadas en el texto
}
\description{
Extractor de sentencias citadas
}
