% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema.R
\name{ccc_tema}
\alias{ccc_tema}
\title{Búscador de sentencias por índice temático}
\usage{
ccc_tema(q, p = 0)
}
\arguments{
\item{q}{términos de búsqueda}

\item{p}{número de página (default es cero)}
}
\value{
un data frame
}
\description{
Búscador de sentencias por índice temático
}
