% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ccc_test}
\alias{ccc_test}
\title{Title}
\usage{
ccc_test(
  fit.l.linear,
  fit.l.logistic,
  fit.r.linear,
  fit.r.logistic,
  contrast,
  lmm_re,
  logmm_re,
  sandwich,
  sender,
  receiver,
  ligand,
  receptor
)
}
\description{
Title
}
