% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar-textos.R
\name{ccc_text}
\alias{ccc_text}
\title{Download Text}
\usage{
ccc_text(url)
}
\arguments{
\item{url}{character string}
}
\value{
text
}
\description{
This function is necessary because many rtf files don't contain footnotes
and many footnotes contain citations.
}
