% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer-texto.R
\name{ccc_texto_pp}
\alias{ccc_texto_pp}
\title{Extractor de sentencia y pies de p치gina}
\usage{
ccc_texto_pp(path)
}
\arguments{
\item{path}{la direcci칩n de la sentencia (e.g. "/relatoria/2006/C-355-06.htm")}
}
\value{
el texto de la sentencia y sus pies de p치gina
}
\description{
Extractor de sentencia y pies de p치gina
}
