% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt-downloads.R
\name{ccc_txt}
\alias{ccc_txt}
\title{Download Text}
\usage{
ccc_txt(url)
}
\arguments{
\item{url}{character string}
}
\value{
text
}
\description{
This function is necessary because many rtf files don't contain footnotes
and many footnotes contain citations.
}
