% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{center_covar}
\alias{center_covar}
\title{Preprocess a data.table: center specified columns (numerical and converted categorical) in place.}
\usage{
center_covar(dt, covar)
}
\arguments{
\item{dt}{A data.table (will be modified directly).}

\item{covar}{A vector of column names to be centered.}
}
\value{
A character vector containing the names of the columns that were centered.

dt <- data.table(
A = factor(c("a", "b", "a")),
B = c(1, 2, 3),
C = c("x", "y", "z"),
D = 4:7,
E = 8:11
)
covar_cols <- c("A", "B", "E")
centered_cols <- preprocess_dt_center_covar_inplace(dt, covar_cols)
print(centered_cols)
print(dt) # The original 'dt' is now modified
}
\description{
Preprocess a data.table: center specified columns (numerical and converted categorical) in place.
}
