% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citations}
\alias{citations}
\title{Citations}
\format{
A data frame with five variables:
\describe{
\item{\code{from}}{Ruling ID (Source)}
\item{\code{to}}{Ruling ID (Target)}
\item{\code{weight}}{Number of times cited}
\item{\code{from_date}}{Source Date}
\item{\code{to_date}}{Target Date}
}
}
\source{
https://www.corteconstitucional.gov.co/relatoria/
}
\usage{
citations
}
\description{
The `citations` dataset contains and edge list of rulings that form a large
citation network.
}
\examples{
  citations
}
\keyword{datasets}
