% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classification_fun}
\alias{classification_fun}
\title{Title Classification for the potential Biomarkers}
\usage{
classification_fun(PBx, newy, method = "RF")
}
\arguments{
\item{PBx}{data matrix}

\item{newy}{New assigned clusters}

\item{method}{Classification method for the function to use}
}
\value{
Predicted values for each observation
}
\description{
Title Classification for the potential Biomarkers
}
\examples{
# classification_fun(PBx, newy, method='RF')
}

