% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster_euclidean}
\alias{cluster_euclidean}
\title{Title Clustering Using Euclidean distances}
\usage{
cluster_euclidean(x, k.gap, plot.clustering)
}
\arguments{
\item{x}{data matrix}

\item{k.gap}{positive integer specifying the number of clusters, less than the number of observation. Default value is 10.}

\item{plot.clustering}{if TRUE (default) a 2-dimensional <U+00E2>Ä<U+0153>clusplot<U+00E2>Äù plot will be printed}
}
\value{
vector with the new assigned clusters
}
\description{
Title Clustering Using Euclidean distances
}
\examples{
# cluster_euclidean(subx,  k.gap = 5, plot.clustering = TRUE)

}

