% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustering}
\alias{clustering}
\title{Title Clustering}
\usage{
clustering(x, k.gap = 2, method = "Euclidean", plot.clustering = TRUE)
}
\arguments{
\item{x}{data matrix}

\item{k.gap}{positive integer specifying the number of clusters, less than the number of observation. Default value is 10.}

\item{method}{Indicating which method to use for clustering. Default is 'Euclidean'.}

\item{plot.clustering}{if TRUE (default) a 2-dimensional “clusplot” plot will be printed}
}
\value{
vector withnew assigned clusters
}
\description{
Title Clustering
}
\examples{
 clustering(subx, k.gap = 5, method='Euclidean', plot.clustering=TRUE)

}
