% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_citation_adj_mat}
\alias{create_citation_adj_mat}
\title{Sparse Citation Matrix}
\usage{
create_citation_adj_mat()
}
\value{
a dgCMatrix
}
\description{
Creates a citation matrix from the edge list in ccc::citations
}
