% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dtm}
\alias{create_dtm}
\title{Sparse Document-Term Matrix}
\usage{
create_dtm()
}
\value{
a dgCMatrix
}
\description{
Creates a document-term matrix from the built-in docterms dataset
}
\examples{

create_dtm()

}
