% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{data.sim}
\alias{data.sim}
\title{Example Dataset}
\format{
\subsection{\code{data.sim}}{

A list with the following elements:
\describe{
\item{\code{counts}}{a matrix of counts with 1000 rows (genes) and 1200 columns (cells)}
\item{\code{metadata}}{a data frame of cell-level metadata with 1200 rows and 4 columns (\code{cell_id}, \code{sample}, \code{group}, and \code{cell_type})}
\item{\code{gene_info}}{a data frame with 1000 rows and 9 columns:
\itemize{
\item \code{gene_id}: simulated gene symbols, in the form of "G" followed by numbers.
\item \code{baseline_log_mu}: baseline log mean count
\item \code{dispersion}: dispersion
\item \code{random_intercept_variance}: variance for sample-level random intercept
\item \code{CT1_logfc}, \code{CT2_logfc}, \code{CT3_logfc}: log fold change of cell-type-specific genes in "CT1", "CT2", and "CT3" cell types
\item \code{grp2_logfc}, \code{grp3_logfc}: log fold change comparing "grp2" and "grp3" to "grp1" group
}
}
}
}
}
\source{
Created using \code{\link[=sim_count]{sim_count()}}.
}
\usage{
data.sim
}
\description{
Example scRNA-seq data from simulation.
}
\keyword{datasets}
