% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docterms}
\alias{docterms}
\title{Document Terms}
\format{
A data frame with three variables:
\describe{
\item{\code{id}}{Ruling ID}
\item{\code{lemma}}{A standardized word}
\item{\code{n}}{Number of times used in the document}
}
}
\source{
https://www.corteconstitucional.gov.co/relatoria/
}
\usage{
docterms
}
\description{
The `doc_terms` dataset contains and edge list of rulings and lemmas (or
processed words).
}
\examples{
  docterms
}
\keyword{datasets}
