% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extract_citations}
\alias{extract_citations}
\title{Extractor de sentencias citadas}
\usage{
extract_citations(texto)
}
\arguments{
\item{texto}{el texto de una sentencia, tal y como viene de ccc_extraer_texto()}
}
\value{
una lista de sentencias citadas (C, SU, o T) en el texto.
}
\description{
Extractor de sentencias citadas
}
