% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FDR_selection}
\alias{FDR_selection}
\title{Title Feature Selection Using FDR selection}
\usage{
FDR_selection(x, y, q = 0.05, eta = 0.1)
}
\arguments{
\item{x}{data matrix}

\item{y}{categorical variable (factor)}

\item{q}{adjusted p value threshold level. The chosen variables will have adjusted p value smaller than q}

\item{eta}{eta squared threshold, the chosen variables will have eta value greater then eta.}
}
\value{
Returns a list of the selected variables
}
\description{
Title Feature Selection Using FDR selection
}
\examples{
# FDR_selection(x, y, q = 0.001, eta = 0.1)
}

