% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{feature_selection}
\alias{feature_selection}
\title{Title Features Selection}
\usage{
feature_selection(x, y, method = "RF", ...)
}
\arguments{
\item{x}{Data matrix}

\item{y}{Dependent variable}

\item{method}{The method to be used for the feature selection: Random forest, AIC, AIC with MSFDR or BIC}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
a list of important variables
}
\description{
Title Features Selection
}
\examples{
# feature_selection(x, y, method='RF')
# feature_selection(x[, 1:30], y, method='BIC')
# feature_selection(x, y, method='FDR_screening')
}
