% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Feature_Selection_AIC}
\alias{Feature_Selection_AIC}
\title{Title Feature Selection Using AIC}
\usage{
Feature_Selection_AIC(x, y)
}
\arguments{
\item{x}{data matrix}

\item{y}{categorical variable (factor)}
}
\value{
Returns a list with two items. The first is a list of important variables. The second
is NA if print.summary.AIC==FALSE or the summary of AIC if TRUE.
}
\description{
Title Feature Selection Using AIC
}
\examples{
# Feature_Selection_AIC(x, y)
}
