% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Feature_Selection_BIC}
\alias{Feature_Selection_BIC}
\title{Title Feature Selection Using BIC}
\usage{
Feature_Selection_BIC(x, y, nbest = 1, nvmax = 12, nmin = 4,
  plot.BIC = FALSE)
}
\arguments{
\item{x}{Data matrix}

\item{y}{response vector (must be numeric?)}

\item{nbest}{number of subsets of each size to record}

\item{nvmax}{maximum size of subsets to examine}

\item{nmin}{number of minimum varibles to be included in the suggested final model}

\item{plot.BIC}{if TRUE (default) the function plots a table of models showing which variables are in each model.
The models are ordered by the specified model selection statistic.}
}
\value{
vector with the names of variables of the model with minimum BIC between the models including more then 'nmin' variables' of regsubsets object
}
\description{
Title Feature Selection Using BIC
}
\examples{
#  Feature_Selection_BIC(x[, 1:30], y, nbest=1, nvmax=5, plot.BIC=TRUE, nmin=4)
}

