% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Feature_Selection_dummy_regressions}
\alias{Feature_Selection_dummy_regressions}
\title{Finds a subset of variables based on all dummy regressions
Title Feature Selection Dummy Regression}
\usage{
Feature_Selection_dummy_regressions(x, y, FUN, ...)
}
\arguments{
\item{x}{Data matrix}

\item{y}{Dependent variable}

\item{FUN}{Indicating which method to use for feature selection}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
a vector with the names of the important variables
}
\description{
Finds a subset of variables based on all dummy regressions
Title Feature Selection Dummy Regression
}
\examples{
Feature_Selection_dummy_regressions(x, y, Feature_Selection_RF)

}

