% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Feature_Selection_RF}
\alias{Feature_Selection_RF}
\title{Title Feature Selection Using Random Forest}
\usage{
Feature_Selection_RF(x, y, p = 0.1)
}
\arguments{
\item{x}{Data matrix}

\item{y}{Categorial dependent variable (factor)}

\item{p}{Precentage of the number of variables to be chosen from x. Default value is 0.1.}
}
\value{
list of p precentage of the variables chosen by their Gini importance index.
}
\description{
Title Feature Selection Using Random Forest
}
\examples{
# Feature_Selection_RF(x, y, p = 0.1)

}

