% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_cases}
\alias{gender_cases}
\title{Gender Equality Cases}
\format{
A data frame with two variables:
\describe{
\item{\code{id}}{Ruling ID (Source)}
\item{\code{tema}}{Topic}
}
}
\source{
https://www.corteconstitucional.gov.co/relatoria/equidaddegenero.php
}
\usage{
gender_cases
}
\description{
Covers cases from 1992-08-12 to 2022-02-23
}
\details{
https://github.com/acastroaraujo/ccc/tree/master/data-raw/gender.R
}
\examples{
  gender_cases
}
\keyword{datasets}
