% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_PBx_from_DATA_C3}
\alias{get_PBx_from_DATA_C3}
\title{Title get_PBx_from_DATA_C3}
\usage{
get_PBx_from_DATA_C3(DATA, META_DATA)
}
\arguments{
\item{DATA}{Full data matrix, includes all observations for all the variables}

\item{META_DATA}{Need to have at least 2 columns, one with all variables name, another one which indicate 
the type of each variable (CM, DX, PB)}
}
\value{
a list of important variables
}
\description{
Title get_PBx_from_DATA_C3
}
\examples{
# PBx <- get_PBx_from_DATA_C3(DATA, META_DATA)

}
