% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_xy_from_DATA_C2}
\alias{get_xy_from_DATA_C2}
\title{Title get_xy_from_DATA_C2}
\usage{
get_xy_from_DATA_C2(DATA, META_DATA)
}
\arguments{
\item{DATA}{Full data matrix, includes all observations for all the variables}

\item{META_DATA}{Need to have at least 2 columns, one with all variables name, another one which indicate 
the type of each variable (CM, DX, PB)}
}
\value{
a list of important variables
}
\description{
Title get_xy_from_DATA_C2
}
\examples{
# x <- get_xy_from_DATA_C2(DATA, META_DATA)[[1]]
# y <- get_xy_from_DATA_C2(DATA, META_DATA)[[2]]
}
