% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hclust_man}
\alias{hclust_man}
\title{Title Deciding on number of clusters by Hierarchical clustering manhattan}
\usage{
hclust_man(x, k.gap, plot.clustering)
}
\arguments{
\item{x}{data matrix}

\item{plot.clustering}{if TRUE (default) a 2-dimensional 'clusplot' plot will be printed}
}
\value{
a list of two variables the hclust function description and a summary table
 of the distribution to clusters
}
\description{
Title Deciding on number of clusters by Hierarchical clustering manhattan
}
\examples{
hclust_man(subx, k.gap = 5, plot.clustering=TRUE)

}

