% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jctt_cases}
\alias{jctt_cases}
\title{Justicia Constitucional en Tiempos de Transicion}
\format{
A chacter vector of rulings
}
\source{
http://justiciatransicional.uniandes.edu.co/web/
}
\usage{
jctt_cases
}
\description{
Covers cases from 1992-10-28 to 2019-04-03
}
\details{
https://github.com/acastroaraujo/ccc/tree/master/data-raw/transitional_justice_uniandes.R
}
\examples{
  jctt_cases
}
\keyword{datasets}
