% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{k_euclidean}
\alias{k_euclidean}
\title{Title Gap statisic with k-medoids euclidean}
\usage{
k_euclidean(x, K.max, B, verbose, plot.num.clus)
}
\arguments{
\item{x}{Data matrix}

\item{K.max}{the maximum number of clusters to consider, must be at least two. Default value is 10.}

\item{B}{integer, number of Monte Carlo (“bootstrap”) samples. Default value is 100.}

\item{verbose}{integer or logical, determining if “progress” output should be printed. The default prints
one bit per bootstrap sample. Default value is FALSE.}

\item{plot.num.clus}{if TRUE (default) the gap statistic plot will be printed}
}
\value{
the clusGap function' values
}
\description{
Title Gap statisic with k-medoids euclidean
}
\examples{
# k_euclidean(subx, K.max=8, B=50, verbose=FALSE, plot.num.clus=TRUE)

}
