% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{k_manhattan}
\alias{k_manhattan}
\title{Title Gap statisic with k-medoids manhattan}
\usage{
k_manhattan(x, K.max, diss, B, cluster.only, verbose, plot.num.clus)
}
\arguments{
\item{x}{data matrix}

\item{K.max}{positive integer specifying the number of clusters, less than the number of observations. 
Default value is 10.}

\item{diss}{if TRUE (default as FALSE) x will be considered as a dissimilarity matrix}

\item{B}{integer, number of Monte Carlo (“bootstrap”) samples. Default value is 100.}

\item{cluster.only}{if true (default) only the clustering will be computed and returned, see details.}

\item{verbose}{integer or logical, determining if “progress” output should be printed. The default prints
one bit per bootstrap sample. Default as FALSE.}

\item{plot.num.clus}{if TRUE (default) the gap statistic plot will be printed}

\item{...}{another objects of pam function}
}
\value{
clusGap function' output
}
\description{
Title Gap statisic with k-medoids manhattan
}
\examples{
#  k_manhattan (subx, K.max = 8, diss=FALSE, B = 50, cluster.only = TRUE, verbose = FALSE)

}
