% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{khclust_euc}
\alias{khclust_euc}
\title{Title  Gap statistics for hclust Euclidean}
\usage{
khclust_euc(x, K.max, B, verbose, plot.num.clus)
}
\arguments{
\item{x}{data matrix}

\item{K.max}{positive integer specifying the number of clusters, less than the number of observations.}

\item{B}{integer, number of Monte Carlo (<U+00E2>Ä<U+0153>bootstrap<U+00E2>Äù) samples}

\item{verbose}{integer or logical, determining if <U+00E2>Ä<U+0153>progress<U+00E2>Äù output should be printed. The default prints
one bit per bootstrap sample}

\item{plot.num.clus}{if TRUE (default) the gap statistic plot will be printed}
}
\value{
the clusGap function output
}
\description{
Title  Gap statistics for hclust Euclidean
}
\examples{
# khclust_euc(subx,K.max=10, B=60, verbose = FALSE, plot.num.clus=TRUE )

}

