% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{khclust_man}
\alias{khclust_man}
\title{Title Gap statistics for hclust Manhattan}
\usage{
khclust_man(x, K.max, B, verbose, plot.num.clus)
}
\arguments{
\item{x}{data matrix}

\item{K.max}{positive integer specifying the number of clusters, less than the number of observations.
Default value is 10}

\item{B}{integer, number of Monte Carlo (“bootstrap”) samples. Default value is 100.}

\item{verbose}{integer or logical, determining if “progress” output should be printed. The default prints
one bit per bootstrap sample. Default value is FALSE.}

\item{plot.num.clus}{if TRUE (default) the gap statistic plot will be printed}
}
\value{
the clusGap function output
}
\description{
Title Gap statistics for hclust Manhattan
}
\examples{
# khclust_man(subx, K.max=8, B=60, verbose=FALSE, plot.num.clus=TRUE)

}
