% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{lr.sim}
\alias{lr.sim}
\title{Example Ligand-Receptor Pairs}
\format{
\subsection{\code{lr.sim}}{

A data frame with 20 rows and 2 columns: \code{ligand} and \code{receptor}.
}
}
\source{
Created using \code{\link[=sim_lr]{sim_lr()}}.
}
\usage{
lr.sim
}
\description{
Example ligand-receptor pairs database from simulation. This is designed to be used in conjunction with \link{data.sim}.
}
\keyword{datasets}
