% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\alias{citations}
\alias{docterms}
\title{30 Year Data}
\source{
https://www.corteconstitucional.gov.co/relatoria/
}
\usage{
metadata

citations

docterms
}
\description{
These datasets go over the first 30 years of the Colombian Constitutional 
Court's history. They were created using this package. See the following 
link to get access to the scripts that produced them (00-05):
}
\details{
https://github.com/acastroaraujo/ccc/tree/master/data-raw

\describe{
  \item{metadata}{A data frame with additional information on each case.}
  \item{citations}{An edge list.}
  \item{docterms}{A data frame with word counts for a subset of the vocabulary used across 
all cases.}
}
}
\examples{
metadata
citations
docterms
}
\keyword{datasets}
