% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MSFDR}
\alias{MSFDR}
\title{Title Forward Selection Using AIC Criteria and MSFDR Procedure}
\usage{
MSFDR(minimal.lm, maximal.lm, q, print.the.steps,
  print.running.time = print.the.steps)
}
\arguments{
\item{minimal.lm}{lm function output of model which includes an intercept}

\item{maximal.lm}{lm function output of model which not includes an intercept}

\item{q}{Significant level. default as 0.05}

\item{print.the.steps}{if TRUE the Lambda, model size, and final model  at each iteration will be printed;
Default as FALSE}

\item{print.running.time}{If TRUE the running time will be printed, it is equal to the value of print.the.steps
Default as False.}
}
\value{
Final model, running time, summary of AIC_MSFDR object
}
\description{
Title Forward Selection Using AIC Criteria and MSFDR Procedure
}
\examples{
# Feature_Selection_AIC_MSFDR(x, y, q = 0.5, print.the.steps = FALSE)

}
