% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{number_of_clusters}
\alias{number_of_clusters}
\title{Title LASSO}
\usage{
number_of_clusters(x, method = "Euclidean", K.max = 10, B = 100,
  verbose = FALSE, plot.num.clus = TRUE, scale = TRUE, diss = FALSE,
  cluster.only = TRUE)
}
\arguments{
\item{x}{Data matrix}

\item{method}{character string indicating how the <U+00E2>Ä<U+0153>optimal<U+00E2>Äù number of clusters: Euclidean (default), Manhattan, 
heirarchical euclidean or heirarchcal manhattan}

\item{K.max}{the maximum number of clusters to consider, must be at least two. Default value is 10.}

\item{B}{integer, number of Monte Carlo (<U+00E2>Ä<U+0153>bootstrap<U+00E2>Äù) samples. Default value is 100.}

\item{verbose}{integer or logical, determining if <U+00E2>Ä<U+0153>progress<U+00E2>Äù output should be printed. The default prints
one bit per bootstrap sample. Default value is FALSE.}

\item{plot.num.clus}{if TRUE (default) the gap statistic plot will be printed}

\item{scale}{if TRUE (default) the data matrix will be scaled.}

\item{diss}{if TRUE (default as FALSE) x will be considered as a dissimilarity matrix.}

\item{cluster.only}{if true (default as FALSE) only the clustering will be computed and returned, see details.}

\item{y}{Dependent variable}

\item{x}{Data matrix}
}
\value{
plot and table which advises how many clusters should be

plot and table which advises how many clusters should be
}
\description{
Title LASSO
}
\examples{
# LASSO_selection(x, y)
Title Deciding on Number of Clusters

}

