% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ccc.R
\name{plot_ccc}
\alias{plot_ccc}
\title{cccのオブジェクトを描画する}
\usage{
plot_ccc(ccc, add = FALSE, xi = "n", pfine = 1, ...)
}
\arguments{
\item{ccc}{cccのオブジェクト}

\item{add}{図に追加するかどうか}

\item{pfine}{錐の明瞭さ。錐の弧の点の数 = 1+pfine*rho(度) (切り上げ)}

\item{...}{polygonのその他引数。色などを指定する}
}
\description{
cccのオブジェクトを描画する
}

