% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramilowski.R
\docType{data}
\name{ramilowski}
\alias{ramilowski}
\title{Ligand-Receptor Table from Ramilowski et al. 2015}
\format{
\subsection{\code{ramilowski}}{

a data frame with 1894 rows and 2 columns: \code{ligand} and \code{receptor}.
}
}
\source{
\url{https://static-content.springer.com/esm/art\%3A10.1038\%2Fncomms8866/MediaObjects/41467_2015_BFncomms8866_MOESM611_ESM.xlsx}
(Supplementary data 2 of the reference paper)
}
\usage{
ramilowski
}
\description{
A data frame of binary ligand-receptor pairs, where each entry is a gene symbol.
Only literature supported ligand-receptor pairs are kept.
}
\references{
Ramilowski, J., Goldberg, T., Harshbarger, J. et al. A draft network of ligand–receptor-mediated multicellular signalling in human. Nat Commun 6, 7866 (2015).
\url{https://doi.org/10.1038/ncomms8866}
}
\keyword{datasets}
