% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RF_classify}
\alias{RF_classify}
\title{Title Classification Using Random Forest Without Down Sampling}
\usage{
RF_classify(PBx, newy)
}
\arguments{
\item{PBx}{data matrix}

\item{newy}{New assigned clusters}
}
\value{
The predicted values for each observation
}
\description{
Title Classification Using Random Forest Without Down Sampling
}
\examples{
# RF_classify(PBx, newy)
}

