% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RF_one_by_one}
\alias{RF_one_by_one}
\title{Title Classification Using Random Forest Without Down Sampling}
\usage{
RF_one_by_one(PBx, newy)
}
\arguments{
\item{PBx}{data matrix}

\item{newy}{New assigned clusters}
}
\value{
a list of two variables: the hclust function description and a summary table
 of the distribution to clusters
}
\description{
Title Classification Using Random Forest Without Down Sampling
}
\examples{
# RF_one_by_one(PBx, newy)
}

