% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_count}
\alias{sim_count}
\title{Simulate Data}
\usage{
sim_count(
  seed = NULL,
  n_genes = 1000,
  n_cell_types = 3,
  cell_type_prob = rep(1, n_cell_types),
  sample_group = c(4, 4, 4),
  cells_per_sample = rep(100, sum(sample_group)),
  re_prop = 0.8,
  re_sigma_shape = 3,
  re_sigma_rate = 2,
  grp_specific_prop = 0.2,
  grp_similarity = 0.5,
  avg_logfc = rep(1, length(sample_group) - 1),
  up_grp_prob = rep(0.5, length(sample_group) - 1),
  ct_specific_prop = 0.05,
  logfc_ct = rep(1, n_cell_types),
  up_ct_prob = rep(1, n_cell_types),
  cell_cont_prop = 0,
  cell_disc_prop = 0,
  sample_cont_prop = 0,
  sample_disc_prop = 0,
  is_cc_confounder = FALSE,
  is_cd_confounder = FALSE,
  is_sc_confounder = FALSE,
  is_sd_confounder = FALSE,
  baseline_log_mu_mean = 1,
  baseline_log_mu_sd = 0.3,
  phi_shape = 2,
  phi_rate = 1,
  lib_size_factor_shape = 3,
  lib_size_factor_rate = 3
)
}
\arguments{
\item{seed}{random seed to use.}

\item{n_genes}{number of genes to simulate.}

\item{n_cell_types}{number of cell types to simulate.}

\item{cell_type_prob}{a vector of length \code{n_cell_types} containing weights for simulating number of cells in each cell type. Need not sum to 1.}

\item{sample_group}{a vector specifying the number of samples in each group. The length of this vector specifies the total number of groups. The first group is considered as the reference group.}

\item{cells_per_sample}{a vector of length \code{sum(sample_group)} specifying the number of cells in each sample.}

\item{re_prop}{numeric specifies the proportion of genes with sample-level random intercepts.}

\item{re_sigma_shape, re_sigma_rate}{Simulate gene-wise standard deviation for sample-level random intercepts from a inverse Gamma distribution with provided shape and rate.}

\item{grp_specific_prop}{numeric specifies the proportions of differentially expressed (DE) genes between each non-reference group and the reference group.}

\item{grp_similarity}{numeric in \link{0, 1} specifies the overlap across the DE genes in each non-reference group.}

\item{avg_logfc}{a vector of length \code{length(sample_group) - 1}. Specifies the average log fold change for each non-reference group compared to the reference group.}

\item{up_grp_prob}{a vector of length \code{length(sample_group) - 1}. Specifies the probabilities of up-regulation in each non-reference group.}

\item{ct_specific_prop}{numeric specifies the proportions of cell-type-specific genes in each cell type.}

\item{logfc_ct}{a vector of length \code{n_cell_types}. Specifies the average log fold change for each cell type.}

\item{up_ct_prob}{a vector of length \code{n_cell_types}. Specifies the probabilities of up-regulation in each cell type.}

\item{cell_cont_prop, cell_disc_prop, sample_cont_prop, sample_disc_prop}{proportions of genes associated with cell-level continuous covariate, cell-level discrete covariate, sample-level continuous covariate, and sample-level discrete covariate respectively. Default to 0.}

\item{is_cc_confounder, is_cd_confounder, is_sc_confounder, is_sd_confounder}{logical specifying whether each covariate and group effects are confounded. Default to FALSE.}

\item{baseline_log_mu_mean, baseline_log_mu_sd}{Simulate baseline log mean for each gene from a normal distribution with provided mean and sd.}

\item{phi_shape, phi_rate}{Simulate the reciprocal of gene-wise dispersion parameter from a Gamma distribution with provided shape and rate.}

\item{lib_size_factor_shape, lib_size_factor_rate}{Simulate library size factors for each cell from a Gamma distribution with provided shape and rate. Recommend these be equal.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{counts}} a matrix of counts, with rows corresponding to genes and columns corresponding to cells.
\item{\code{metadata}} a data frame containing cell-level metadata (e.g., cell type, group, sample, covariates).
\item{\code{gene_info}} a data frame containing, for each gene, its baseline log mean, dispersion, within-sample correlation, and log fold change.
}
}
\description{
Simulate multi-sample multi-group single-cell RNA-seq data using negative binomial distribution.
}
\examples{
# Simulate some data using the default
dat <- sim_count(seed = 1)

dat$counts[1:6, 1:6]
head(dat$metadata)
head(dat$gene_info)

}
