% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_lr}
\alias{sim_lr}
\title{Simulate Ligand-Receptor Pairs}
\usage{
sim_lr(
  seed = NULL,
  genes.l = paste0("G", 1:500),
  genes.r = paste0("G", 501:1000),
  dimer = 0.25,
  trimer = 0.1,
  n_lr = 100
)
}
\arguments{
\item{seed}{random seed to use.}

\item{genes.l, genes.r}{vectors specify genes as potential ligands and receptors respectively.}

\item{dimer}{numeric. Probability of ligand/receptor consisting of two subunits.}

\item{trimer}{numeric. Probability of ligand/receptor consisting of three subunits.}

\item{n_lr}{desired number of ligand-receptor pairs.}
}
\value{
description
}
\description{
Simulate ligand-receptor pairs from given gene lists, taking into account subunit architecture.
}
\examples{
# Simulate some ligand-receptor pairs using the default
lr <- sim_lr(seed = 2)
head(lr)

# Simulate some binary ligand-receptor pairs
lr.bin <- sim_lr(seed = 3, dimer = 0, trimer = 0)
head(lr.bin)

}
