% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_make.R
\name{all_make}
\alias{all_make}
\title{Undescribed}
\usage{
all_make(samples, coneN, height, azimuth, control)
}
\arguments{
\item{samples}{Sample}

\item{coneN}{The number of sample}

\item{height}{The size of cone}

\item{azimuth}{The lower angle of cone}

\item{control}{options}
}
\description{
Undescribed
}
\examples{
samples <- data.frame(x = rnorm(10), y = rnorm(10))
dat <- all_make(samples, 100, 1, pi/2, control = list(max_loop = 1e+05))
}

