% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_analysis.R
\name{ccc_analysis}
\alias{ccc_analysis}
\title{Title}
\usage{
ccc_analysis(
  expression_matrix,
  metadata,
  contrast,
  cell_id_col = "cell_id",
  cell_type_col = "cell_type",
  group_col = "group",
  covar_col = NULL,
  cdr = TRUE,
  id_col = NULL,
  lmm_re = TRUE,
  logmm_re = TRUE,
  sender = NULL,
  receiver = NULL,
  lr = c("omnipathr", "ramilowski"),
  multi_sub = c("minimum", "arithmetic_mean", "geometric_mean", "min_avg_gene",
    "min_rate_gene"),
  sandwich = FALSE,
  verbose = TRUE,
  min_pct = 0.01,
  large_n = 2,
  min_avg_pct = 0,
  min_cell = 10,
  threshold = 0,
  sep_prop = 0,
  sep_n = 0,
  sep_detection = TRUE,
  padj_method = "BH",
  cell_type_padj = TRUE,
  control_logm = list(),
  control_lmm = lme4::lmerControl(),
  control_logmm = list(),
  chunk_size = 10,
  nthreads = 1
)
}
\arguments{
\item{multi_sub}{it 
\itemize{
  \item
  \item
}

@param cell_type_padj adjust p-values for each sender-receiver pair or not}
}
\description{
Title
}
