% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer_cosas.R
\name{ccc_pp}
\alias{ccc_pp}
\title{Extractor de pie de páginas}
\usage{
ccc_pp(link)
}
\arguments{
\item{link}{(e.g. "/relatoria/2006/C-355-06.htm")}
}
\value{
una lista de pies de página
}
\description{
Extractor de pie de páginas
}
