% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraer-texto.R
\name{ccc_texto}
\alias{ccc_texto}
\title{Extractor de sentencia}
\usage{
ccc_texto(path)
}
\arguments{
\item{path}{la dirección de la sentencia (e.g. "/relatoria/2006/C-355-06.htm")}
}
\value{
el texto de la sentencia
}
\description{
Extractor de sentencia
}
