% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender}
\alias{gender}
\title{Gender Equality Cases}
\source{
https://www.corteconstitucional.gov.co/relatoria/equidaddegenero.php
}
\usage{
gender
}
\description{
Gender Equality Cases
}
\keyword{datasets}
