% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.normalize.R
\name{log_normalize}
\alias{log_normalize}
\title{Log Normalize Data}
\usage{
log_normalize(count_matrix, scale_factor = 10000, base = exp(1))
}
\arguments{
\item{count_matrix}{a numeric matrix of raw counts, with rows corresponding to genes and columns corresponding to cells.}

\item{scale_factor}{numeric. The scale factor fro cell-level normalization. Defaults to 10000.}

\item{base}{numeric. The base of the log transformation. Defaults to \code{exp(1)}.}
}
\value{
A numeric matrix of normalized counts.
}
\description{
Log normalize the count matrix. Gene counts for each cell are divided by the total counts for that cell and multiplied by \code{scale.factor}. This is incremented by 1 and then log transformed.
}
\examples{
set.seed(123)
raw <- matrix(rnbinom(100, mu = 1.5, size = 2), nrow = 20)
normalized <- log_normalize(raw)
normalized

}
