% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata}
\format{
A data frame with nine variables:
\describe{
\item{\code{id}}{Ruling ID}
\item{\code{type}}{Ruling Type (C, T, or SU)}
\item{\code{year}}{Year}
\item{\code{date}}{YYYY-MM-DD}
\item{\code{indegree}}{Citation In-degree}
\item{\code{outdegree}}{Citation Out-degree}
\item{\code{word_count}}{Word Count}
\item{\code{descriptors}}{Additional metadata added to some rulings by the Court}
\item{\code{url}}{URL to see the ruling}
}
}
\source{
https://www.corteconstitucional.gov.co/relatoria/
}
\usage{
metadata
}
\description{
The `metadata` dataset contains information about the rulings authored by
the Colombian Constitutional Court.
}
\examples{
  metadata
}
\keyword{datasets}
