% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ccc.R
\name{plot_ccc}
\alias{plot_ccc}
\title{Plot ccc}
\usage{
plot_ccc(ccc, add = FALSE, xi = "n", pfine = 1, ...)
}
\arguments{
\item{ccc}{ccc's object}

\item{add}{If TRUE, it added the current picture. By default, FALSE.}

\item{pfine}{Delicacy of cone. delicacy = 1+pfine*rho(degree) (rounded up)}

\item{...}{The other options}
}
\description{
Plot ccc
}

